requirements.txt
scipion-pyworkflow
scipion-em